finished combinging into DecompTree
